refactor(settings): 移动默认设置和配置接口到独立模块，优化代码结构 #57
